move api auth out of internal
